Extract common base class
